Wfp 1126 (#129)
* WFP-1126 get assessments from WMT staging tables

* WFP-1126 lint fixed

* WFP-1126 removed redundant column mappings

* WFP-1126 refactor

* WFP-1126 refactor calculator for CMS workload points

* WFP-1126 added find contacts outside caseload

* WFP-1126 added find contacts in caseload performed by others

* WFP-1126 added getting reductions in real time for overview, refactored

* WFP-1126 changed reduction setup to start of day to match wmt-web

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsworkload/domain/OffenderManagerOverview.kt
@@ -22,7 +22,7 @@ data class OffenderManagerOverview @JsonCreator constructor(
   @Schema(description = "Team Name", example = "Team Name")
   val teamName: String,
   @Schema(description = "Probation Practitioner total cases", example = "25")
-  val totalCases: BigDecimal,
+  val totalCases: Int,
   @Schema(description = "Probation Practitioner Contracted hours per week", example = "37")
   val weeklyHours: BigDecimal,
   @Schema(description = "Probation Practitioner total Reduction hours per week", example = "10")
